feat: admin-new url -> admin 으로 폴더명 변경
