--- app/admin-new/page.tsx
@@ -1,10 +1,0 @@
-import React from "react";
-
-import Admin from "./Admin";
-import "@/style/reset.css";
-
-function AdminPage() {
-  return <Admin />;
-}
-
-export default AdminPage;
