+++ app/admin-old/page.tsx
@@ -1,0 +1,6 @@
+import Home from "@/home/Home";
+// import Login from "@/login/Login";
+
+export default function adminLoginPage() {
+  return <Home />;
+}
