+++ app/admin/(components)/ContentArea.tsx
