+++ app/admin/(components)/ThemeDrawer/helpers/imageHelpers.ts
