+++ app/admin/(components)/ThemeInfo/ThemeInfoHint.tsx
